\docType{data}
\name{USArrests}
\alias{USArrest}
\title{USArrests}
\format{The data used here refers to crimes and arrests in USA.
\describe{
This data set contains statistics, in arrests per 100,000 residents for assault, murder, and rape in each of the 50 US states in 1973. Also given is the percent of the population living in urban areas.)
}}
\source{
World Almanac and Book of facts 1975. (Crime rates).

}
\usage{
USArrests
}
\description{

}
\keyword{datasets}
